+++ madprops/serializers.py
@@ -170,6 +170,19 @@ class PropertySerializer(ModelSerializer):
         return data
 
 
+class NestedPropertySerializer(PropertySerializer):
+    """Base class for Nested Property class."""
+    @property
+    def fields(self):
+        # We need this to skip validation of foreign key field if properties
+        # are created together with properties owner. In this case parent
+        # field value is not known on a property validation stage, so
+        # validation fails.
+        fields = super(PropertySerializer, self).fields
+        fields.pop(self.opts.parent_obj_field, None)
+        return fields
+
+
 class PropertiesOwnerSerializer(ModelSerializer):
     """Base class for "parent" serializers.